SDFV: Adds a separate embedded layout
This layout is a preparation for embedding the SDFV in space-limited
containers, where having the moving sidebar disturbs usability. In
this layout, the sidebar is replaced with an info-box below the
canvas, which dynamically changes content in a similar way to the
sidebar in the traditional layout. The change is made in such a way
that it strikes a balance between being minimally-invasive in the
traditional layout, while avoiding code duplication for the embedded
layout.